Change the approach to escaping symbols
Previously `Tokenizer` consumed all the backslashes and it was impossible to understand whether a character was escaped or not in later stages of parsing.
Now I introduced class `CmdChar`, which is basicaly a char with a boolean flag, whether it is escaped or not.
I also introduced class `CmdString` and `CmdStringBuilder` which are prety much like the classes from the standard library, but `CmdString` consists of `CmdChar`s.